refactor(dashboard): extract orchestra-drawer from orchestra.ts
Split orchestra.ts (842 LOC) into orchestra.ts (448) + orchestra-drawer.ts (427) + orchestra-signals.ts (7).
Drawer/detail components, SVG sub-components, bounds/camera computation, and shared signals extracted to reduce main file below 500 lines.
